Improve pruning lower bound in bounded search functions
For `for_paths_with_area_and_bounce`, this change speeds up pruning
enormously. I conjecture that that is because it improves the branching
factor at the outer levels of the search, which has a large impact on
the overall speed, and having `bounce_min` as a starting point means it
is not too expensive to perform the search.

For `for_paths_with_area`, the equivalent change slows things down. I
tried several variations, and somehow changing from the closed-form
division to an iterator search is the fastest of all.