Switch to one-dimensional maps for table calculation
It was using a two-dimensional map object because I developed that while
working on an entirely different implementation strategy (that didn't
pan out) that specifically used a two-dimensional map, but it turns out
that the one-dimensional one is a bit faster somehow. It kind of seems
like it ought to be the same, but I'm definitely seeing a difference.