Clear maps directly at top level
I thought I was being clever and reducing the need for reallocation with
the old thing, but it turns out that doubles memory usage while helping
speed marginally at best. The speed gain does seem to be real, and it's
just enough that it might be worth putting in a switch, but let's keep
it simple for now.